requirements/requirements-test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements/requirements-test.in
#
alabaster==0.7.13
    # via sphinx
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
    #   django-stubs
attrs==24.3.0
    # via
    #   outcome
    #   trio
babel==2.16.0
    # via sphinx
backports-zoneinfo==0.2.1
    # via
    #   django
    #   djangorestframework
build==1.2.2.post1
    # via pip-tools
certifi==2024.12.14
    # via
    #   requests
    #   selenium
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via pip-tools
colorama==0.4.6
    # via sphinx-autobuild
cryptography==44.0.0
    # via
    #   jwt
    #   oic
defusedxml==0.7.1
    # via oic
distlib==0.3.9
    # via virtualenv
django==4.2.18
    # via
    #   -r requirements/requirements.in
    #   django-cors-headers
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-cors-headers==4.4.0
    # via -r requirements/requirements-test.in
django-stubs[compatible-mypy]==5.1.0
    # via
    #   -r requirements/requirements-dev.in
    #   djangorestframework-stubs
django-stubs-ext==5.1.2
    # via django-stubs
djangorestframework==3.15.2
    # via -r requirements/requirements-test.in
djangorestframework-stubs[compatible-mypy]==3.15.1
    # via -r requirements/requirements-dev.in
docutils==0.19
    # via
    #   sphinx
    #   sphinx-rtd-theme
exceptiongroup==1.2.2
    # via
    #   trio
    #   trio-websocket
filelock==3.16.1
    # via virtualenv
future==1.0.0
    # via pyjwkest
h11==0.14.0
    # via wsproto
identify==2.6.1
    # via pre-commit
idna==3.10
    # via
    #   requests
    #   trio
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.5.0
    # via
    #   build
    #   sphinx
isort==5.13.2
    # via -r requirements/requirements-test.in
jinja2==3.1.5
    # via sphinx
jsonpickle==4.0.1
    # via -r requirements/requirements.in
jwt==1.3.1
    # via -r requirements/requirements.in
livereload==2.7.1
    # via sphinx-autobuild
mako==1.3.8
    # via oic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mypy==1.11.2
    # via
    #   -r requirements/requirements-dev.in
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
oic==1.7.0
    # via -r requirements/requirements.in
outcome==1.3.0.post0
    # via trio
packaging==24.2
    # via
    #   build
    #   sphinx
pip-tools==7.4.1
    # via -r requirements/requirements-dev.in
platformdirs==4.3.6
    # via virtualenv
pre-commit==3.5.0
    # via
    #   -r requirements/requirements-dev.in
    #   -r requirements/requirements-test.in
psycopg2==2.9.10
    # via -r requirements/requirements-test.in
pycparser==2.22
    # via cffi
pycryptodomex==3.21.0
    # via
    #   -r requirements/requirements.in
    #   oic
    #   pyjwkest
pydantic==2.10.5
    # via pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via oic
pygments==2.19.1
    # via sphinx
pyjwkest==1.4.2
    # via oic
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pysocks==1.7.1
    # via urllib3
python-decouple==3.8
    # via -r requirements/requirements-test.in
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.2
    # via babel
pyyaml==6.0.2
    # via pre-commit
requests==2.32.3
    # via
    #   djangorestframework-stubs
    #   oic
    #   pyjwkest
    #   sphinx
ruff==0.9.1
    # via -r requirements/requirements-dev.in
selenium==4.27.1
    # via -r requirements/requirements-test.in
six==1.17.0
    # via pyjwkest
sniffio==1.3.1
    # via trio
snowballstemmer==2.2.0
    # via sphinx
sortedcontainers==2.4.0
    # via trio
sphinx==6.2.1
    # via
    #   -r requirements/requirements-test.in
    #   sphinx-autobuild
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-autobuild==2021.3.14
    # via -r requirements/requirements-test.in
sphinx-rtd-theme==3.0.2
    # via -r requirements/requirements-test.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.5.3
    # via django
tomli==2.2.1
    # via
    #   build
    #   django-stubs
    #   mypy
    #   pip-tools
tornado==6.4.2
    # via livereload
trio==0.27.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
types-docutils==0.21.0.20241128
    # via types-pygments
types-markdown==3.7.0.20241204
    # via -r requirements/requirements-dev.in
types-psycopg2==2.9.21.20241019
    # via -r requirements/requirements-dev.in
types-pygments==2.19.0.20250107
    # via -r requirements/requirements-dev.in
types-pyyaml==6.0.12.20241230
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.0.20241016
    # via djangorestframework-stubs
types-setuptools==75.8.0.20250110
    # via types-pygments
typing-extensions==4.12.2
    # via
    #   annotated-types
    #   asgiref
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
    #   pydantic
    #   pydantic-core
    #   selenium
urllib3[socks]==2.2.3
    # via
    #   requests
    #   selenium
    #   types-requests
virtualenv==20.29.0
    # via pre-commit
websocket-client==1.8.0
    # via selenium
wheel==0.45.1
    # via pip-tools
wsproto==1.2.0
    # via trio-websocket
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
